As reported directly from George Hartwig, Jr.:
"In the late 70's Dr Bruce Henriksen and Dr Stephen Wolff were trying to bring BRL into the world of multi-user computing and as a part of this process bought a PDP 11/70 computer, a variety of terminals and Mike Muuss to install a version of JHU Unix.  Early in this experiment Mike wrote a plotting package called Terminal Independent Graphics so that things could be plotted to terminals, primarily the HP2648A, and printers, again primarily Versatexs. This package was used as part of a chain of programs that were 'piped' together."

Basically, special thanks go to Dr. Wolff and Dr. Henriksen for enabling Mike to get started on BRL-CAD, providing resources and support presumably from a managerial level in the very very early days of BRL-CAD.